requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-telegram-bot==21.3
sqlalchemy==2.0.31
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.0
